# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive OAuth 2.0 service with multi-tenancy, Google OAuth, and service-to-service authentication.",
        "title": "Multi-Tenant OAuth Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/oauth/google/callback": {
            "get": {
                "description": "Handle Google OAuth callback and exchange code for tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Google OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"4/0AfJohXn...\"",
                        "description": "Authorization code from Google",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"tenant-123\"",
                        "description": "State parameter",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"access_denied\"",
                        "description": "Error from Google OAuth",
                        "name": "error",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/google/login": {
            "get": {
                "description": "Redirect user to Google OAuth for authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Initiate Google OAuth login",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"tenant-123\"",
                        "description": "Tenant ID",
                        "name": "tenant_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"https://app.example.com/callback\"",
                        "description": "Custom redirect URI",
                        "name": "redirect_uri",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/one-time": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a one-time token for secure access to specific URLs with session and page restrictions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Create one-time authentication token",
                "parameters": [
                    {
                        "description": "One-time token parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.OneTimeTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.OneTimeTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/token": {
            "post": {
                "description": "Exchange client credentials for access token (service-to-service authentication)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "OAuth 2.0 Client Credentials Flow",
                "parameters": [
                    {
                        "description": "Client credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.ClientCredentialsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ClientCredentialsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/verify": {
            "get": {
                "description": "Verify and consume a one-time token with session and page validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Verify one-time authentication token",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"one-time-token-123\"",
                        "description": "One-time token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"survey-session-456\"",
                        "description": "Session ID",
                        "name": "session",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"survey-submit-page\"",
                        "description": "Page ID",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.OneTimeSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rest.ClientCredentialsRequest": {
            "type": "object",
            "required": [
                "client_id",
                "client_secret",
                "grant_type"
            ],
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "service-client-123"
                },
                "client_secret": {
                    "type": "string",
                    "example": "service-secret-456"
                },
                "grant_type": {
                    "type": "string",
                    "example": "client_credentials"
                },
                "scope": {
                    "type": "string",
                    "example": "read:users write:users"
                }
            }
        },
        "rest.ClientCredentialsResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "scope": {
                    "type": "string",
                    "example": "read:users write:users"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request data"
                }
            }
        },
...
        "rest.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                },
                "user": {
                    "$ref": "#/definitions/rest.User"
                }
            }
        },
        "rest.OneTimeSessionResponse": {
            "type": "object",
            "properties": {
                "is_valid": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Token validated successfully"
                },
                "page_id": {
                    "type": "string",
                    "example": "survey-submit-page"
                },
                "session_id": {
                    "type": "string",
                    "example": "survey-session-456"
                },
                "token": {
                    "type": "string",
                    "example": "one-time-token-123"
                },
                "user": {
                    "$ref": "#/definitions/rest.User"
                }
            }
        },
        "rest.OneTimeTokenRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "max_uses": {
                    "type": "integer",
                    "example": 1
                },
                "page_id": {
                    "description": "Specific page/action identifier",
                    "type": "string",
                    "example": "survey-submit-page"
                },
                "session_id": {
                    "description": "Unique session identifier",
                    "type": "string",
                    "example": "survey-session-456"
                },
                "url": {
                    "type": "string",
                    "example": "https://app.example.com/survey/123"
                }
            }
        },
        "rest.OneTimeTokenResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "max_uses": {
                    "type": "integer",
                    "example": 1
                },
                "page_id": {
                    "type": "string",
                    "example": "survey-submit-page"
                },
                "session_id": {
                    "type": "string",
                    "example": "survey-session-456"
                },
                "token": {
                    "type": "string",
                    "example": "one-time-token-123"
                },
                "url": {
                    "type": "string",
                    "example": "https://app.example.com/survey/123?token=one-time-token-123\u0026session=survey-session-456\u0026page=survey-submit-page"
                }
            }
        },
...
        "rest.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "role-123"
                },
                "name": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "rest.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "user-123"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.Role"
                    }
                },
                "tenant_id": {
                    "type": "string",
                    "example": "tenant-123"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2ClientCredentials": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "/api/v1/oauth/token",
            "scopes": {
                "admin": "\"Admin access\"",
                "read": "\"Read access\"",
                "write": "\"Write access\""
            }
        }
    },
...
}